# lib/api-zod/src/generated/types/movie.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Movie {
  id: number;
  title: string;
  synopsis: string;
  genre: string;
  year: number;
  rating: number;
  durationMinutes: number;
  posterUrl: string;
  backdropUrl: string;
  cast: string[];
  director: string;
  isFeatured: boolean;
  inWatchlist: boolean;
}